Use pythreejs Scene controller to hold scene translation
Currently this uses a workaround that freezes the z component of the
translation given in the constructor in order to have mouse rotation
be reasonable in the resulting three.js scene. A better solution would
be to set the default plato.draw.Scene translation to (0, 0, 0), but
that would potentially require minor changes to other backends.